+++ thoth-app/src/component/work_status_modal.rs
@@ -13,7 +13,8 @@ pub struct Props {
     pub onsubmit: Callback<()>,
     pub oncancel: Callback<()>,
     pub object_name: String,
-    pub current_user: AccountDetails,
+    pub object_work_status: String,
+    pub object_current_work_status: String,
     pub current_state_unpublished: bool,
     pub is_published: bool,
     #[prop_or_default]
@@ -67,9 +68,9 @@ impl Component for ConfirmWorkStatusComponent {
                         </header>
                         <section class="modal-card-body">
                             <p>
-                                { "Are you sure you want to change the work status to Active for " }
+                                { "Are you sure you want to change the work status to" } { &ctx.props().object_work_status  } { " for " }
                                 <i>{ &ctx.props().object_name }</i>
-                                { "? Once a Work has been set to Active, it cannot be set back to Forthcoming." }
+                                { "? Once a Work has been set to" } { &ctx.props().object_work_status }  { ", it cannot be set back to " } { &ctx.props().object_current_work_status }  { "." }
                             </p>
                         </section>